+++ lib/clients/ai/openai/gpt.interface.ts
